refactored contact resource and added AWS SES helper for sending emails
