+++ config/main.ini
@@ -1,4 +1,5 @@
 [DEFAULT]
+aws-region = us-east-1
 id-character-length = 36
 local-base-url = https://opendemic.ngrok.io
 webhook_token = 37d961ad-da4f-4921-915e-c84f62ade295
